Feature/27 buy and sell (#36)
* get balance

* Allow for multiple RSI calculators

* finetune rsi

* fix test

* buy and sell managers

* use sell and buy managers

* add notifications

* Update Trader.cs

* Better group notifications

* add simulation mode

* Update TradeService.cs

* Update TradeService.cs

* logo

* Update KrieptoBot.Application.csproj

* Fixed nuget version issues

* Update SellManager.cs

* Update TradingContext.cs

* Cleanup

* Updated solution version

* Update TradeService.cs

* Increase RSI accuracy

+++ KrieptoBot.Application/Recommendators/RecommendatorRsi14_Period_4h.cs
@@ -5,8 +5,8 @@ namespace KrieptoBot.Application.Recommendators
 {
     public class RecommendatorRsi14Period4H : RecommendatorRsiBase
     {
-        protected override float SellRecommendationWeight => 2F;
-        protected override float BuyRecommendationWeight => 2F;
+        protected override decimal SellRecommendationWeight => 2;
+        protected override decimal BuyRecommendationWeight => 2;
 
         public RecommendatorRsi14Period4H(IExchangeService exchangeService, IRsi rsiIndicator,
             ITradingContext tradingContext, ILogger<RecommendatorRsi14Period4H> logger) : base(exchangeService,
