+++ KrieptoBot.Application/Recommendators/RecommendatorScore.cs
@@ -2,15 +2,15 @@
 {
     public class RecommendatorScore
     {
-        public float Score { get; set; }
+        public decimal Score { get; set; }
 
-        public static RecommendatorScore operator *(RecommendatorScore recommendatorScore, float b) =>
+        public static RecommendatorScore operator *(RecommendatorScore recommendatorScore, decimal b) =>
             new RecommendatorScore { Score = recommendatorScore.Score * b };
-        public static RecommendatorScore operator /(RecommendatorScore recommendatorScore, float b) =>
+        public static RecommendatorScore operator /(RecommendatorScore recommendatorScore, decimal b) =>
             new RecommendatorScore { Score = recommendatorScore.Score / b };
-        public static RecommendatorScore operator +(RecommendatorScore recommendatorScore, float b) =>
+        public static RecommendatorScore operator +(RecommendatorScore recommendatorScore, decimal b) =>
             new RecommendatorScore { Score = recommendatorScore.Score + b };
-        public static RecommendatorScore operator -(RecommendatorScore recommendatorScore, float b) =>
+        public static RecommendatorScore operator -(RecommendatorScore recommendatorScore, decimal b) =>
             new RecommendatorScore { Score = recommendatorScore.Score - b };
         public static RecommendatorScore operator *(RecommendatorScore recommendatorScoreA, RecommendatorScore recommendatorScoreB) =>
             new RecommendatorScore { Score = recommendatorScoreA.Score * recommendatorScoreB.Score };
